[frontend/backend] wip import/export
