+++ opencti-platform/opencti-front/src/schema/relay.schema.graphql
@@ -9028,6 +9028,7 @@ type Mutation {
   fintelTemplateAdd(input: FintelTemplateAddInput!): FintelTemplate
   fintelTemplateDelete(id: ID!): ID
   fintelTemplateFieldPatch(id: ID!, input: [EditInput!]!): FintelTemplate
+  fintelTemplateImport(file: Upload!): String!
   disseminationListSend(input: DisseminationListSendInput!): Boolean
 }
 
@@ -12547,6 +12548,7 @@ type FintelTemplate implements BasicObject & InternalObject {
   template_content: String!
   start_date: DateTime
   fintel_template_widgets: [FintelTemplateWidget!]!
+  configuration_export: String!
 }
 
 input FintelTemplateWidgetAddInput {
